ocean-eye/ai_service/requirements.txt
Flask
requests
transformers
torch
Pillow
opencv-python
exifread
numpy